Fixes, documentation, and unimplemented callbacks
I made a few fixes, most noticeable in the source code (or rather the
other thing) is that it now accounts for the null terminator. Hasn't
been tested yet, but bear with me, it will.

The most exciting update (for me) is the documentation and unimplemented
callback functions, which helps guide me to eventually completing this.

I added a section to describe the predefined macros, which are passed to
the parameter parameter (lol), and examples on some on how they're used.
Also added 2 new macros, FIRST and LAST, description in header.

Added and moved around a ton of callback functions, a lot of which are
not implemented yet. The exciting bit is that, thanks to C's String.h,
majority of the functions can be done with relative ease.