More functions, new features
Added quite a few more functions, which will soon be given documentation
for as soon as all of the others are finished, but the most prominent
addition is the usage of bit masking to smuggle in some cool parameters.

Also changed macro values (and added one) to make the masking possible.
The bit masking now allows you (A.K.A just me) to pass multiple
parameters with the bitwise 'OR' operator, and it looks something like
this.

```
String_Utils s = String_Utils_Create();
char *string = s->concat("Hello", "dlroW ", NO_MODIFY | REVERSE);
printf("%s", string);
```

The above should concatenate the second string in reverse to the first,
returning "Hello World".  When dealing with string literals, its advised
to use the NO_MODIFY flag to avoid runtime errors.