Fixes
Fixed a few functions, and all unit tests pass (except one, which I know
works but don't feel like bothering to check it there, might change that
later).

Even if this is a relatively small project, I have to say it definitely
is exhausting. Tweaking the logic, finding off-by-one errors which stop
everything to a grinding halt, thinking you fixed a bug when it spawns
more bugs or it really was just being hidden by the compiler...
Exhausting, but so going to be worth it in the end, I can feel it!