More additions
Added a few functions and some new callbacks which will implemented as
soon as I get to it. It's almost ready to be moved to Stable section!

The work is seeming to get a bit tedious, but I'll thank myself for this
later on down the line when I use this in another project which saves me
the amount of time by using these function calls. After I added the
other new functions, which I seem to be adding more every day, but not
this time.

After I finish implementing the functions, adding them all to the
constructor (Like 40 god damn callbacks), Unit Testing, THEN
documenting, it'll be stable enough to use in ANY project I used in the
future. Thank goodness for that.