Almost finished!
Heavy Unit testing caused me to make a bunch of changes, A.K.A Fixes,
for problems I would not have known existed otherwise. Thank the Gods
for Unit Tests!  Also added a few functions, and am almost complete all
of the tests for my functions. I will have two Unit Tests in the end,
one calling the normal functions, another which is the exact same except
it calls the Struct's callbacks.

When I am finished with this, it should be stable enough to include in
any project I or anyone else works on.