Me toying around with creating a simple Garbage Collector for my String_Utils, all thanks to: http://journal.stuffwithstuff.com/2013/12/08/babys-first-garbage-collector/
It was interesting to work on, and I vaguely have a  better
understanding of garbage collection, but no where near enough to
implement a simple one, even for my own String_Utils, specializing
solely on Strings. My initial expectations were the following:

"Oh, cool, I use a garbage collector in the case that I don't want to
end up worrying about allocating and deallocating memory all over the
place!" but I walked away with less than expected.

My thoughts on it though, it's neat, but it won't work for my program
as it is, I'd have to give more thought as to how I can tailor it to
my program. It uses a stack and a linked-list of objects which holds
the string's value. Now, this sounds great, in fact, it's wonderful,
because due to mark and sweep, it will free anything not on the
stack... the bad part is that it's on a stack. Lets say I want to
return a copy of a string? Guess what, I have to push the temporary
string on the stack, then pop it off at the end of the function. Read:
I may as well malloc then free myself. Worst yet, not only that, but
since it's a stack, it pops off the very last object added to the
stack... so if I allocate the temporary string and return it, then pop
it off the stack later, it will free that string even though it's in
use.

I'll have to put this off to the side for now, very interesting
though. Maybe, no, definitely after I actually finish String_Utils,
Networking_Utils and File_Utils, AND Data_Structures (and Algorithms?)
as I already have a huge amount on my plate. Much later.