Finished up documentation.
Also another surprise, it actually compiles perfectly in NetBeans, which
is amazing since before it worked in one but not the other. I'm rather
impressed with it actually, never thought I'd do it, but this time I
actually finished my very first library (1.0 was so awful can't even
call it a finished library).

Added a bunch of documentation as well as gotchas, hopefully DoxyGen
picks up on this.