Started on the template for my Linked List. However, this Linked List is no ordinary Linked List! This Linked List will be the base for any other Linked List based implementations I'll be doing. I already see it now.
The Linked List is unimplemented, but I am working on the header files
and brainstorming as to how I should create it. When finished, it should
be generic enough to directly use with little modification for anything.
It has a barebones iterator (struct), the Linked List (struct) has basic
information like it's current size, the pointer to the first and last
nodes, and even doubles as a single and double linked list through a
tagged union.

Now, how exactly is this going to work for any data structure I attempt
to implement? Well for one, it has the option of passing in callback
functions in the constructor, which if set will allow for custom
operations, I.E Delete (Can't really fully delete a nested void * if I
don't know in advance what's inside it now do I?), Add (In the case that
special insertions are needed, like a binary tree), and Compare (For the
eventual sorting function I'll implement, most likely merge sort). What
makes it even more special, is that these aren't needed at all. If you
pass in NULL for those parameters, the default, generic implementations
of them will be used instead, making using it for simple applications
viable.

My mindset right now is to do all of the hard work now so I don't have
to do it later. An extreme of DRY, a naive outlook of YAGNI (who
wouldn't need a fully generic linked list?), and a somewhat lack of KISS
(It's not overly complicated, but it needs to be some complex for what
I'm trying to do). Wish me luck!