Revised the Linked  List header file. This is makes things 100x easier, and honestly I don't see the need to have a single linked list when a double linked list has only slightly more overhead (of course I know that single linked lists are more efficient in some cases, but if I want to make one, I'll do it at a later date).
Changing the whole idea of the Data Structures project too. While I plan
on having all data structures derive from Linked List (as in they will
use and build on top of it), I can't make it generic enough that
literally any type of data structure can work on it. For instance, while
I can make a normally single-linked list implementation like a stack
work with double linked list, a *tree will not work at all, hence I will
have to create a tree list some time in the future... or maybe expand on
this and end up with what I was originally aiming for, except with a
stable data structure to base it on. Regardless, I'm taking it easy this
time.

The Linked List will, when stable, be able to do all of the following:

1) Add and Delete items from the linked list.
1A) Have parameters which will tailor this for you. For instance, there
are already macro definitions which will do certain things based on what
parameters you pass. Like adding DELETE to Linked_List_Remove will
delete the item as well as remove the pointer from the list, by calling
the delete callback (either default or user provided) on it before
finishing up it's operation.
2) Obtain items from the list based on either an iterator or by
supplying the index.
2A) There will be index checking, so if the index given is bogus, it
will either return NULL, or return the first or last (maybe both if
there's only one item) in the list. Haven't decided which yet.
2B) Iterator will be able to also add and remove items from the linked
list, except with more ease than normally accessing the Linked List. It
also has parameter-passing.
3) Ability to sort the linked list based on a comparator passed.
3A) If no comparator passed, it will NOP and return immediately.
3B) Have parameters to either be in descending or ascending order.

And that's all she wrote folks. Hopefully will be able to finish this
within a week or two, but with finals coming up, probably not until
after summer begins. Then it'll be V. 1.0.