Adding some features and implementing as many as I can for now. Already finished up the Linked List (besides for it's destructor). Even started on the iterator. Problem is that I haven't really compiled it, so it's most likely full of bugs, syntax errors (bad part of using a text editor to write code I suppose), and memory leaks, not to mention general logic errors.
I do, however, plan on fixing them as they come. As most likely I am
echoing to myself only (or hopefully future potential employers), I
cannot wait until I finish this linked list. I will take a bit of a
break from it to work on some concurrency. I want to work on my own
thread pool and pretty much an easy and abstract way to create threads
through my own library/API. Of course I'll be using pthreads, but it'll
be a great way to learn how it works, and I'm excited to be able to just
use one or two methods to create my very own worker thread to run in the
background as I do other stuff. HooAh!

Current projects so far seems like this...

1) Data_Structures/Linked_Lists
2) Concurrency_Utils/Thread_Pool
3) Concurrency_Utils/Threads
4) Data_Structures/Hash_Map
5) Data_Structures/Binary_Tree <- For a memory heap.
6) String_Utils/String_Utils_Garbage_Collector <- implement heap
...

Currently finished project (as in, not touching for a long while)

1) String_Utils/String_Utils

Can't wait until I get more done. It's rather fun to learn all of this!