Implemented a few new functions and usability, still haven't compiled it yet, but most likely I've got a huge surprise waiting for me in multitudes of errors, bugs and off-by-ones.
Writing this has become kind of like my own programmer's diary, so here
are a few thoughts...

Projects like this definitely make me thank the Gods that current
implementations exist in other programming languages, and that things
are abstracted to the point where I don't even have to know how a linked
list, array list, hash map or anything works in order to use... but this
project I'm undergoing has become rather enlightening.

But I digress. The main thing I want to talk about (well... to myself I
suppose) is that this project shows me just how much of programming is
actually implementing new ideas and features... or rather, how little.

I swear, there are like 5 - 6 phases of development (at least that I've
experienced).

1) Brainstorming and the Drawing Board:
This is the phase where you begin thinking about what you're going to
do, how you're going to do it, what constraints you have, tools you must
work with, etc. This is the most exciting time for me. Thinking about
how I'm going to implement 'X, Y, and Z' feature, how cool it's going to
be to work on, how I can finish it within a week if I try hard enough...
it's a nice motivational push, that's for sure.

2) Creating the interface... or rather, pre-development:
This is when you are creating the header file for the project, thinking
about how you're going to make this thing work. You begin taking ideas
you had while brainstorming, and then try to implement them as best you
can. This does not always work out, unfortunately like with my case of
attempting to create a Double, Single and Tree list all into one.
Usually during this time you are also making it more difficult for
yourself because you think "Oh, I can do this, should only take a few
minutes" where in reality it takes a few hours. All in fun though.

3) Implementing the interface, A.K.A development:
When you are attempting to follow the header file you create and
implement those functions... then realize, "Hey wait, that won't work"
and have to go back and change the header as you implement the source.
In the end, half of the pre-development options you made are removed and
trimmed to make the project do-able, or modified/optimized to do better.
Then, the worst part, when things you figured were easy weren't actually
as easy as you first thought. You scrap the overly complicated ideas to
make it simpler and more do-able. You stop for the day, come back the
next day and wonder "What the hell was I doing here?" and have to revise
bits and pieces of code you already did. Eventually you manage to finish
it, you compile it, but it doesn't run. Hours to a day later, it finally
compiles... now to the next step.

4) Unit Testing...:
Unit Testing is good, I will admit it, it is extremely necessary and no
project is complete without good tests to ensure that all functions work
as expected. However, writing the test can feel very much like a drag.
Take my String_Utils test, a lot of tests were rather tedious as it was
pretty much testing whether strings were properly manipulated, however
it will definitely bring out 99% of bugs with your program. It's just a
pain having to write it out yourself mostly from scratch. This process
is rather draining.

4A) Static Analysis Tools:
Tools like Valgrind are great for this, and help you discover any leaks
and logic errors/flaws you may have overlooked in your test. This isn't
as bad as the Unit Tests, as they work in conjunction with the Unit
Tests, but some errors can be a pain to track down and properly dispose
of. Once you pass both of these, it's stable enough to use.

5) Documentation... the devil in the details.
The devil truly is in writing out documentation for functions. "They're
your own functions, how bad can it be?" you ask? True I do know what
each does, inside out, but having to write up what everything does, EVEN
if the purpose is already clear in the naming convention of the
variables and functions you use, you still have to write it out. This
devil, depending on the size of the project, can go from hours to a day.
I guess I take many breaks between, but still it is a rather long
process.

5A) Documentation Generation Tools:
Doxygen is a great tool without a doubt, however at times it can be
frustrating. It misses a lot of comments and you have to set flags in
the option just right. Sometimes comments are misaligned and they end up
going to the function or variable after the one you intended. Not too
much of a big deal, but things add up...

I suppose for a 6th you have moving everything to stable, but that's the
easiest of all parts. These 5 things however definitely are things I
have to deal with when programming that I had to vent. Thanks for
listeninig, me.