Decided to add the report anyway, even though it isn't really stable yet. Some good news, the garbage collector was garbage (still is, needs massive improvement... and a heap), but the GCC attribute is just plain amazing. It does what I want it to do, and much more. It deletes the string I mark as TEMP in the declaration, works wonderfully. I can see combining these two aspects together.
One problem I can forsee. clean_up only applies to the variable
declared, not the string it's pointing to, meaning if the string
actually changes to another, it will not retain that value when it
exits the scope. However, honestly, that migh tbe something I'd have
to just deal with by using the set method, which would do it just fine
anyway.

The size of the program actually was cut by 1/3. It was 12 - 13k
before, now it's only 4k (and barely at that). I don't know if the
performance test is accurate, but the library functions don't seem to
do anything that's bad performance-wise, however I still have trouble
printing each value when they're deleted, but ah well, I'll mark it
as a bug and move on. Still have to finish the Linked List.