Began to refractor the Linked_List. Honestly, it's a darn good thing I'm doing this too, can't imagine an employer would be happy to hear it took me over a week to complete a linked list, even if this is my very first full fledged one I am making.
The Linked_List no longer has an iterator, or rather, it no longer has
an iterator struct. Instead, it in itself is an iterator. Also, moved
around some things in the header file so the typedef function pointers
are all grouped together based on what they do. So basically, all I did
today was documentation, and a bit of coding. Ah well, I still cannot
wait for the finished result since, regardless of if there are better,
this one is something I can call mine.