Did quite a bit of refactoring and revision of code. It looks, like, 10x better than before. The static (A.K.A Private) functions look to be coming along well and look good too, well documentation too if I can say so myself.
As they say, if a function/method does more than one thing, then it's
doing it wrong, right? So, I removed the parameter passing to it for
lazy deletion, instead it will do one thing and one thing well.

node_to_index: Returns the index of the node found. In the future, will
have it just return the last or first if they match to save quite a bit
of work.

index_to_node: Will return NULL if index is out of bounds, and the first
and last nodes in the list respectively, and will also in the future
iterate through ->next or ->prev depending on size / 2, and whether or
not the index if greater than or equal to, or less than that. Only if
the size is greater than, say, 10.

item_to_node: Used to be item_to_index or may as well have been called
that. Now it will return the node that the item is located at, making it
easier to pass the node directly to other function that require a node
instead of an item.

node_exists: Like node_to_index, except since 0 is a valid index, the
return values of this function will be 1 if it exists, 0 if it does not
to avoid any logic confusion.

Those should be it, truth be told, and I can finally start to work on
the iterator portion tomorrow. Darn finals hogging up my time, didn't
even get any studying done though.