Create File_Utils.h
Instead of studying for my finals, had to get this out of my head. It's basically a base template for the File_Utils library I'll be making. It pretty much wraps the File in it's own object, of course will be adding function pointers to simulate methods and general ease of use. However, as I wrote this on a library computer, it's a start. The struct contains the file it is being read from, it reads all contents from the file, and even keeps track of the current line number and current character in the line. Barebones and minimal, but it should work.