I know, finals coming up, should be studying, but I felt it couldn't wait. Actually, during my finals I had even more ideas and couldn't concentrate, hope that doesn't reflect too badly on my grades.
Added a bunch of functions, even a summary of what it will do when finished. I'm making it how I would want to manipulate and read files while working with the constraints of using C. In essence, File_Utils basically creates a copy of the file by storing each new line into a an array of buffers, each holding up to the next newline, regardless of the size. Honestly, the entire, file could be on gigantic runon sentence and it wouldn't know, nor do I intend it to. This way it will be possible to flush the contents back to file again without any loss of the original format. I mean, technically I could write everything in to a 1kb buffer each, and just keep the newlines of each, but the whole point is that reads up to the next line. Imagine how much difficult it'd be having to parse each file for the newline, then concatenate it to a temporary buffer if the newline was in the next one or two buffers.

I.E

Start at buffer[0], next newline is at buffer[2], then I'd have to concatenate buffer 0 and 1, then 1 and 2 up to the newline, and keep track of that as well. I'll consider that later when I get more invested in the HOW, I'm more invested in the WHAT and WHY.

Anyway, this allows it to create a copy of the original file, and even overwrite the original file. Problem with it is that normal operations like appending to a huge file would take a long time anyway, since it has to open the file, store the contents and parse each newline into an array of buffers, and then the user has to call Append_To_End which, is an O(1) operation, but having to copy it for such a simple task... it won't be efficient. I'll end up creating a simpler function for appending to a file though.