Added a bunch of macros which vastly increase the readability of the source code, and I mean vastly. Get_Task, which was the only one I was working on today (besides for changing the necessary LoC to be macro-fied in others) is almost literally as readable as any comment or documentation. Self-Documenting code, indeed.
The macros are basically to shorten the pthread_* prefix on every function, but I also added a macro to atomically increment and decrement variables such as thread_count and active_threads. Needless to say, it does make it more readable, and in my opinion, elegant.

Also, I think I have the basics of Get_Task done, but haven't run it yet, but it looks to be fine, hopefully I can have it up and running by the end of the week (taking my jolly old time, few hours a day of coding). Wish me luck!