Redesigned the repository by removing Concurrency Utils entirely, it's not replaced with Thread Pool, which, as you guessed it, just contains the Thread Pool. The concurrent data structures are going to be in, you guessed it, Data Structures.
The Thread Pool is not fully tested yet, however I will do that later. I will however release it as 'stable' as the original test works, and I am confident I can get the timed functions to work, but I don't have the time to test it yet, as it's gonna be a busy weekend.