Note to Self: Add a NU_Client_connect function which works in a reusable way.
Also consider using a Ring Buffer.

The reasoning for separating create and connect is because of reusability; it normally would be impossible if I did it the other way, connecting during creation, causing the user to have to destroy the client, then create it again, wasting overhead. Also, having a basic creation method would allow my NU_Server_t to spawn clients from the given sockets withotu having to worry about creating a smaller, private function to spawn it, or connecting to some random arbitrary host just to receive the client that's connected to it.

Also on a side note, a ring buffer sounds very jolly and all, and would help me add that to the list of available data structures... but how could I make it, internally, thread safe? I know I can externally lock the use of it, but that confines it's use within one function (safely). While a normal char * buffer can easily be created and destroyed at will, overhead be damned.

Anyway, I'm taking my time once again, so not many too major updates, tomorrow I should have client finished and be working on the ring buffer, but this project might not be done for a week or two, like the Thread Pool.

Once again, it's not that I'm moving too slow, but that this is my own tempo. I program for 2 - 3, maybe 4 - 5 on some days, hours at a time and then cool down doing something else.