Kind of at a loss here as to whether I should use the bounded buffer for writing as well as reading, but that'd feel redundant. In sending a message, it involves storing the message in a buffer of some kind, and the act of copying, or worse, assigning it to the bounded buffer's buffer, would take away the entire point. Even if you argue that a string literal could easily be used, what if they're using gets to read from a stream, like stdin or a file? Then you need a buffer anyway!
At least in receive, the bounded buffer can be used as it's abstracted to the point that the user doesn't need to worry about passing their own. The client's bounded_buffer is created when the client is, hence the only allocations needed are at creation, unless it's resized; however, if the passed buffer_size to be read is of the same size as the previous, there is no reallocation needed.

Another note, resize_buffer sets the index in the case that the buffer is reused without needing to clear the entirety of the buffer. I might end up removing that portion, however I'm not sure yet.