Implemented the 5 functions I marked as TODO earlier, they are as follows:
bsock_to_string, clients_to_string and data_to_string return a string representation of the structs in a formatted way, although asprintf may be a bit of a problem for the for loops, but nothing else I can do.

delete_all_clients and delete_all_sockets do as they say, or should anyway. By moving up the head, then re-assinging the recently freed reference back to the head should be the easiest way to queue up deletions.