In MU_Logger, added MU_LOG_ASSERT, which may replace MU_ASSERT_RETURN entirely, as returning after logging isn't really that difficult, but it did allow me to conditionally return, but I'm trying to make it more readable by making everything inside of blocks rather than the one-liners I used to write.
Second, I succumbed to the need to use a goto. Oh the horror! The deallocation portions for allocating 2 resource pools for server, was becoming extremely long winded. The goto in this type of case seems to be good practice, as it's even used in the Linux kernel, and for jolly good reason too. Now, instead of deallocating right there, I log the error/assert and then goto error.