Alright, a few things today. For one, I made a more generic makefile, but not quite. It has to be specific because I only wish to compile certain things. I'll try to create some .mk files to make it easier to maintain multiple make files, but unfortunately this one took me 3 - 4 hours because I tried to use everything that the guide mentions, %.o: %.c, and didn't know that VPATH was necessary. Hence, I might have been able to make it more generic. *Sigh* ah well..
Another thing: I began working on the Formatter, I managed to get it
running in IDE, or rather a barebones one. It worked by having a string
passed to it with the format %fnc and the like, and it could actually
successfully do so. I even used hashes to make switch statements
possible, so it'd be slightly faster than strcmp. All in all, good
enough.