A relatively large update. I implemented my new MU_ARG_CHECK and MU_RETRY libraries, which ARG_CHECK does as it says, checks arguments (and unfortunately due to stinginess of stringification, it unrolls NULL back to it's actual macro, ((void *)0), so I leave it out) and MU_RETRY attempts to implement what TEMP_FAILURE_RETRY did by wrapping the macro in a while loop, except it takes a storage argument of which it will store the return of the function. The bad part? It won't work on void ones, but those are relatively few as is, I could always create a separate one just for those exceptions.
Besides that, yeah, I didn't break NU_Server, as the test runs
perfectly, and I even found a way to make it reliably show the website
reliably; apparently if I do not accept the header, the browser
considers the transaction as "incomplete" so closing the server leads to
a ERR_CONN_RESET on Chrome, and god knows what on others.

I feel rather content though, the new logging looks beautiful, could
even remove the hard-coded function names in the error logs, as it now
gets the function that calls it in the log. 100% better.

After this, I'll definitely have to make Client work again, then it's on
to HTTP, then I can even test out my DS_Hash_Map data structure.