DS_Hash_Map now has the option to change the max key size at compile via use of the preprocessor and compiling with the -D flag, with a default of 128 bytes in size.
NU_HTTP is also following suit in allowing the user to determine the
max size of the header fields and whatnot, with an optional default too
of course.

Also, I'm starting on implementing the HTTP parser. Going smoothly.

Also in DS_Hash_Map, I have the keys allocated on the stack, so constant
read-only keys can be used as well, and even temporary buffered keys can
be added to be freed immediately after since the key in each bucket is
allocated on the struct's stack.