I kinda "finished" it, but I'm not sure, I still am not 100% sure how it works, but I get the gist of it.
Basically what it does it, you can "acquire" or claim access to data,
upon which prevents it from being deleted until no other thread
currently holds the data. When it is released, it basically gets added
to the retirement list, and as long as one of the threads in the hazard
pointer list maintains a pointer to it, it will not be deleted. This is
the extent at which how I think it works, it feels like a mark-sweep
algorithm except a LOT more tricky.

I gotta try it out though, dunno how, but I'll figure a way, most likely
via implementing that stack.