Added a BUNCH of aliases to begin on the promise I made in the README. Personally, I 100% think this is needed. The style difference alone can make people want to look for another, and I agree that it definitely is different, but macro 100%, while significantly increasing collision, makes it more readable and fit in with everything else.
I.E DS_Stack_create(...) is kind of verbose, as if it was C++, it'd be DS::Stack::create. However, the cstd definitely has things like stack_create(...), which fits in wonderfully with everything
else. Only issue is that everyone and their mother has implemented a stack in their life time, and hence would cause major collisions. However, that's what the macro is for, it's optional.

I also completed documentation for DS_Hash_Map which I found was actually incomplete.