I created the ref_count wrapper I spoke of earlier, although I have no tests for it, the example I got the original idea from seems to work, but the issue isn't obtaining the meta data, the issue unfortunately is whether the atomic reference counter is good enough to ensure thread safety.
I also began implementing event_loop_fd more, first as static functions,
and later as individual functions. Looking at it though, it 100% should
be capable of handling any async task which doesn't bog down the thread.
The issue with reading files, just like Android, is that it CAN be slow,
but delegating to another thread helps alleviate this, and to help even
more, the user can submit event_sources as write, read, or read-write.
This means that if you asynchronously read from the event_loop, you can
then add another source to write that data as well, asynchronously.
Granted, it won't be as pretty, but oh well.