Deleted all past logs as they are irrelevant now.
I also changed the internal node correction algorithm, as I realized
that prev->next and next->prev are extremely unreliable. For example,
imagine the sequence 1->100->1000. Now, 100 gets removed and 2 is now
inserted into a sorted list...

1->2->1000

Now, our current node is the one pointing to 100. Since 100 is no longer
valid, we go to next, or 1000. Note that, even if the new sequence is...

1->2->3->...->1000

it's fine to skip to 1000 since that originally was the next item in the
list before we became invalidated. While it may be an outdated ordering
of the sequence, it is a correct one none-the-less. Now, lets say 1000
is removed and it is

1->2->3->...999

Now, next is no longer valid. Since it is no longer valid, we go back to
prev, or 1. So, now we get prev->next, which is 2.

This would yield results such as....

1->100->2

Which is awful. Hence, we only check if next (or prev) is valid and skip
to that. Otherwise, we are invalidated and start over. This way we
yield...

1->100

Which is fine, in comparison. Once again, if you want ALL of it
reliably, you can use iterator_for_each or list_for_each.

Lastly, besides for updating event_loop, I decided to add conf to stack
to test how it would look if lock_free was an optional feature. It looks
good, so I'll continue on with that.