Added a bit more to map, not much today, did mostly research. Apparently the memcmp will not work for most structs (ones not allocated with calloc or had been runthrough with memset), as the padding would differ even if all members were actually the same.
However, it seems that what I can do instead is to serialize it into an
array of bytes, and then hash that ahead of time. Of course, that would
be up to the caller (although I will add an extra configuration callback
to automatically apply this transformation for you), but it makes it
possible for the caller to handle hashing structs on their own. Also, if
the caller HAD properly zero'd the struct ahead of time, it would work
the same, or if they decide to pack instead.