I just realized something really big... Really big...
Due to the inexperience I've had with general algorithms, I naively
implemented a blocking SORTED LIST, not a blocking priority queue. For
one thing, the blocking_queue as of yet is VERY susceptable to
starvation, and also was lazily implemented probably becuase I got tired
midway through writing it.

Anyway, apparently I should impelment this using a heap. It works for
now, but for production code, it is awful. It functions as a normal
blocking queue, but makes for an awful priroity queue, which just
reflects badly on me... *Sigh* I have soo much crap to do right now, 3
exams this week (FML), and now this project is seriously burning me out
fast. I might have to just come back to it at another date like before,
but I'd rather finish it. We'll see.

For now, I have to implement a Binary Tree which can function as a Heap,
then have the priority queue use that.