I began to create the tree, but almost nothing to done. What I like about this is that I'm definitely learnign something completely and entirely new. Never created a tree before, so making a binary heap is definitely something new and fresh.
The list of things needing to be done so far...

1) Test the hash map implementation, and implement it's iterator.
2) Create buffer_t to allow for easier serialization of structs and fix
up the serialize.h helper macros
3) Make Blocking_Queue use a binary heap... of course gotta create that
first.
4) Thread Pool's priority queue needs to not starve the hell out of it's
work queue. This could use a comparator which not only compares based on
priority but also time inserted.
5) event_t needs to be able to wait on multiple events
6) event_loop_fd needs to be actually completed, as many other features
rely on it... oh wait, that also relies on socket_utils.h...
7) Fix up http.h because it's a massive piece of shit. Also won't work
anymore until map.c is reimplemented correctly (FML)
8) Create socket_utils.h, future_t, etc. God help me...
9) ADD CONFIGURATION SUPPORT TO EVERY GOD DAMN OBJECT IN THIS LIBRARY,
that's like what 40 files?
10) Actually TEST the reference counting (so god help me if there's an
oversight in it's implementation)

Needless to say, those are one of the top things I'm having planned. I'm
going to have to promise myself to not add more and more features.
Getting out of hand now...