Made quite a few advancements. For one, adding a maximum size to list.c was extremely easy, all I had to do was add another configuration option for it's configuration object, and boom, thats it. Doesn't break anything else.
I also finished the heap and tested it, it works! Yay! I also began
using it in the blocking_queue, which is going to be using a normal list
for normal blocking queue operations, and the heap for prioritized
blocking queue operations. This way obtaining the max should be O(1) and
O(log(N)) respectively, and inserting a new task should be O(1) and
O(log(N)) respectively.

Its cool though, blocking_queue pretty much wraps an existing data
structure and handles it that way. I need to complete some more
implementation stuff, but enqueue and dequeue are finished. The rest is
easy. Another thing, the underlying list/heap also handle deleting and
removing everything from the list. They can handle decrementing the
reference counts, or invoking the necessary destructor, etc.